third_party/WebKit/LayoutTests/http/tests/security/contentTypeOptions/block-audio-as-script.html

<!DOCTYPE html>
<head>
    <title>Block 'audio/*' when served as script.</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
      window.scriptsSuccessfullyLoaded = 0;
    </script>
</head>
<body>
    <script src="resources/script-with-header.pl?mime=audio/aiff&amp;options=invalid"></script>
    <script src="resources/script-with-header.pl?mime=AudiO/midi&amp;options=invalid"></script>
    <script src="resources/script-with-header.pl?mime=AUDIO/wav&amp;options=invalid"></script>
    <script src="resources/script-with-header.pl?mime=audio/x-wav&amp;options=invalid"></script>
    <script src="resources/script-with-header.pl?mime=audio/whatever+xml&amp;options=invalid"></script>
    <script>
      test(function () {
        assert_equals(window.scriptsSuccessfullyLoaded, 0);
      }, "All audio/* scripts should be blocked.");
    </script>
</body>
</html>
